+++ scripts/train.py
@@ -2,13 +2,13 @@
 
 import gzip
 import logging
-import multiprocessing as mp
 import sys
 from pathlib import Path
 from typing import Optional, TextIO
 
 import torch
 import torch.distributed as dist
+import torch.multiprocessing as mp
 import wandb
 from packaging import version
 from torch.distributed.fsdp import FullyShardedDataParallel as FSDP
@@ -124,6 +124,7 @@ def main(cfg: TrainConfig) -> None:
     # Wrap the model in FSDP.
     log.info("Wrapping model with FDSP...")
     wrap_policy = olmo_model.get_fsdp_wrap_policy(cfg.fsdp.wrapping_strategy)
+
     if version.parse(torch.__version__) >= version.parse("2.1.0"):
         # This prevents any parameters from being initialized twice
         def dummy_init_fn(module: torch.nn.Module) -> None:
@@ -241,11 +242,17 @@ def main(cfg: TrainConfig) -> None:
 
 
 if __name__ == "__main__":
-    mp.set_start_method("spawn")
+    try:
+        mp.set_start_method("spawn", force=True)
+    except RuntimeError as e:
+        print(f"failed to set multiprocessing start method: {e}")
+
     # Initialize process group.
     dist.init_process_group(backend="nccl")
 
     prepare_cli_environment()
+
+    log.info(f"multiprocessing start method set to '{mp.get_start_method()}'")
 
     try:
         yaml_path, args_list = sys.argv[1], sys.argv[2:]